+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v0.2.1
 
 - In default scopes, replace offline per offline_access (The standard value)
+- Add `--ttl` and `--renewAt` parameters on `client` and `client-nui` subcommand to exercise token renewal.